refactor: move search functionality to separate SearchBar component
